wintun: stop recreating and deleting TUN interface
This commits reuses the current TUN interface. Instead of creating and
deleting it each time, it only enables it when in use and disables it
when not in use.

- This allows reuse of existing interface for WHLK testing.
- This allows adapter association with a NLA profile to persist across
  connections.
- This allows user configured network bindings to persist  across
  connections.

Do not use in production! Without flushing the adapter to vanilla state
on each connection this is a potential security risk.

Do not use with wireguard-windows client: a MSI custom action is
required to clean all TUN adapters before uninstalling.

Signed-off-by: Simon Rozman <simon@rozman.si>